pyproject.toml
[project]
name = "orgahome"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "starlette>=0.50.0",
  "uvicorn[standard]>=0.38.0",
  "python-dotenv>=1.2.1",
  "aiohttp>=3.13.2",
  "authlib>=1.6.5",
  "jinja2>=3.1.6",
  "aiofiles>=25.1.0",
  "httpx>=0.28.1",
  "itsdangerous>=2.2.0",
]

[dependency-groups]
dev = ["ruff>=0.14.8", "ty>=0.0.1a32"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
orgahome = "orgahome:cli"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["F401"]
